All sigma-point transform--based filters (UKF, GHKF, CKF, ...) now inherit from SigmaPointTransform class which defined common apply() method. (I was tired of code duplication). This fixed the bug appearing when sigmas are more than 1-D.
Fixed __init__() of the TPQKF, hypers need to be (n, ) array not (n, 1).
Added some tests, added imports to the __init__() of the inference package, so that when from inference import * is called all the filters are imported.

EKF still does not work on non-additive UNGM. Gives posdef errors, need to check Jacobians and look more closely at the non-additive version of the Taylor-based transform.