check_jacobians() works for both noise cases. EKF, CKF and UKF fail the tests for non-additive UNGM due to 1-th leading minor not positive definite errors.
Strange that GaussHermiteKalman, GPQuadKalman, TPQuadKalman  are passing the test, so I don't think this has to do with the sigma-point transform code or the StateSpaceInference code. Could be that the system is simply such that these filters inevitably fail. Try on pendulum or find better example of nonadditive noise system.