Unit sigma-point and hype-parameter initialization pulled out of GPQuadKalman's and TPQuadKalman's __init__() methods. Changed tests to supply defaults.
The filter classes should probably supply defaults in case the unit SPs and hypers are not supplied. Starting to think that it might the responsibility for defaul unit SP and hypers settings should be delegated to the BQ transforms. This would be best handled by defining another parent class BayesianQuadratureTransform, which would define apply() and set_defaults() methods.