Derived and coded weights for the cross-covariance. Adjusted computation of Q, which now results in positive average model variance. The transformation is thus complete.
The problem is, that the filter with the UT points (derivative in the middle) and default hypers diverges. The computation of the derivative covariances Kdd and Kfd still gives badly conditioned matrices. This causes the kernel matrix K to be badly conditioned which subsequently affects it's inverse iK. Should have a hard look at computation of Kdd.