Fix StudentianInference.
Tested on the experiments in tpq_ungm.py and tpq_constant_velocity.py. The UNGM yields very similar results to those in the TPQ paper.

GaussianInference uses the new method get_moments() to obtain the mean and covariance of the initial/noise RVs.

Improve docstrings: first line is one line summary, below it is a more detailed description.