Fix TPQ experiments.
The experiments are not completely fixed yet. The UNGM works more or less. Doesn't quite yield the same conclusion as the TPQ paper, but it's close.

The constant velocity with glint noise however is not working. The Student filter (SF) with fully-symmetric set diverges and so are the TPQSFs, albeit slower. The UKF works fine.