Fix: old Student code: GP and TP expected model variance.
EMV is now returned as vector, in case of (single-output) TP, and as a number, in case of (single-output) GP. The quadratic form in the TP case was replaced with an np.einsum fix.

Add: parameter dim_out to GPQ and TPQ transforms.

The impact on the overall RMSE and INC  is negligible.